Add direct method to using packages in index.md
As a newcomer to the Javascript ecosystem, tools like Grunt, RequireJS,
Yeoman and so on are entirely new to me, and so it's unhelpful to be
expected to know what these tools do before understanding how to use
the packages installed by Bower.

This is the simplest way to use Bower (although, I assume, not the
best), I think it makes it clear that all Bower does is install packages
in `bower_components` and it is up to the user to use them or build them
as needed.